+++ crates/ludtwig-parser/CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT-VERSION
+
+# v0.5.0
 - Fix HTML Tag name parser token collisions, which caused tags like `source` to not parse correctly
 - Fix Twig accessor parser to allow array indexing by dot operator
 - Fix lexer and parsing of colon word syntax like `:special-attribute` as HTML attribute and now support `array[:upper]` as Twig slice